+++ src/com/iriszero/ILogDatabase.java
@@ -1,0 +1,10 @@
+package com.iriszero;
+
+import java.util.List;
+
+public interface ILogDatabase
+{
+    List<UserMessage> GetLast(long n);
+    List<UserMessage> GetAll();
+    void Add(UserMessage msg);
+}
