+++ src/com/iriszero/MySqlDatabase.java
@@ -1,0 +1,9 @@
+package com.iriszero;
+
+public class MySqlDatabase
+{
+    MySqlDatabase(String user, String password, String url)
+    {
+
+    }
+}
